Made a Helper Class 04/27/16
I made a helper class.
