+++ nbproject/private/private.xml
@@ -3,7 +3,7 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/C:/Users/Set/Desktop/Desktop/School/CS/3rd%20yr/2nd%20Sem/CS%20124/Project/DungeonFactory/src/dungeonfactory/DungeonFactory.java</file>
+            <file>file:/C:/Users/mnmkami/Documents/GitHub/DungeonFactory/src/dungeonfactory/Game.java</file>
         </group>
     </open-files>
 </project-private>
